drf-docker-react-nginx/backend/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.8
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
webencodings==0.5.1
psycopg2-binary
